– long instead of int for the epoch values

+++ app/src/main/java/cz/uruba/ets2mpcompanion/constants/Numeric.java
@@ -1,5 +1,5 @@
 package cz.uruba.ets2mpcompanion.constants;
 
 public class Numeric {
-    public static final int gameTimeStartingPointEpoch = 1445731200;
+    public static final long gameTimeStartingPointEpoch = 1445731200;
 }
